LayoutTests/fast/dom/DOMURL/invalid-url-getters-expected.txt

Test what getters return on an invalid URL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS invalidURL.toString() is 'http://@@:p@www.example.com:22/bar?x#y'
FAIL invalidURL.origin should be . Was null.
PASS invalidURL.username is ''
PASS invalidURL.password is ''
PASS invalidURL.host is ''
PASS invalidURL.hostname is ''
PASS invalidURL.port is ''
PASS invalidURL.pathname is ''
PASS invalidURL.search is ''
PASS invalidURL.hash is ''
PASS successfullyParsed is true

TEST COMPLETE
